Change BlobWriter to return an URL
I am not sure if the way the azure blob writer assigns URLs is good. But
it is possible that a blob provider might store blobs under a name that
si not the cid, e.g. the underlying storage does not allow assigning free
names, or there are some constraints regarding directory size.

So we need the ability for the azure provider to assign and return an url.